% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{addcountries}
\alias{addcountries}
\title{addcountries()
This function add a "countries" column in the dataframe by deducing the location of earthquake using long and lat.}
\usage{
addcountries(df)
}
\arguments{
\item{df}{takes a dataframe}
}
\value{
a dataframe
}
\description{
addcountries()
This function add a "countries" column in the dataframe by deducing the location of earthquake using long and lat.
}
\examples{
\dontrun{
addcountries(MyData)
}
}
