% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{dayfunction}
\alias{dayfunction}
\title{dayfunction()
This function subset the data to the number of days prior to current time inserted my the user input}
\usage{
dayfunction(x, df)
}
\arguments{
\item{x}{takes an integer. Number of days prior to current time}

\item{df}{takes a dataframe}
}
\value{
a dataframe
}
\description{
dayfunction()
This function subset the data to the number of days prior to current time inserted my the user input
}
\examples{
\dontrun{
dayfunction(3, MyData) #data from 3days ago till current time.
}
}
