% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{earthquakedepth}
\alias{earthquakedepth}
\title{earthquakedepth()
This function returns a dataframe in corellation with the earthquake depth input by user.}
\usage{
earthquakedepth(x, df)
}
\arguments{
\item{x}{takes object input. Only accept LOW, INTERMEDIATE, and DEEP as answer}

\item{df}{takes a dataframe}
}
\value{
a dataframe
}
\description{
earthquakedepth()
This function returns a dataframe in corellation with the earthquake depth input by user.
}
\examples{
\dontrun{
earthquake(deep, MyData)
}
}
