% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{hourfunction}
\alias{hourfunction}
\title{hourfunction()
This function subset the data to the number of hours prior to current time inserted my the user input}
\usage{
hourfunction(x, df)
}
\arguments{
\item{x}{takes an integer. Number of hours from current time}

\item{df}{takes a dataframe}
}
\value{
a dataframe
}
\description{
hourfunction()
This function subset the data to the number of hours prior to current time inserted my the user input
}
\examples{
MyData = importusgsdata(day) #importing dataframe
hourfunction(25, MyData) #data from 25 hours ago till current time.
}
