% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{importusgsdata}
\alias{importusgsdata}
\title{importusgsdata()
This function imports live-feed data from USGS. The User have a choice to choose live-feed data for the past hour, day, week, month}
\usage{
importusgsdata(charac, df = FALSE)
}
\arguments{
\item{charac}{Takes an object called HOUR, DAY, WEEK, or MONTH.}

\item{df}{User choose to export dataframe automatically to the global environment called MyData. Defaults to FALSE.}
}
\value{
a dataframe
}
\description{
importusgsdata()
This function imports live-feed data from USGS. The User have a choice to choose live-feed data for the past hour, day, week, month
}
\examples{
\dontrun{
importusgsdata(week, df = TRUE)
importusdsdata(hour)
}
}
