% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{magnitude}
\alias{magnitude}
\title{magnitude()}
\usage{
magnitude(lowbound = NULL, upbound = NULL, df)
}
\arguments{
\item{lowbound}{takes a float or int. Defaults to NULL}

\item{upbound}{takes a float or int. Defaults to NULL}

\item{df}{takes a dataframe}
}
\value{
a dataframe
}
\description{
This function takes a dataframe and also the lowerbound and upperbound from user input and produce a dataframe for magnitude ranges between lowerbound and upperbound.
}
\examples{
\dontrun{
magnitude(2,5,MyData)
magnitude(upbound = 3, MyData)
}
}
