% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{mapfunction}
\alias{mapfunction}
\title{mapfunction()
This function maps the location of earthquake onto a leaflet map. This map function is an interactive map that the user can user and interect to see their data in a geospatial manner.}
\usage{
mapfunction(df)
}
\arguments{
\item{df}{takes a dataframe}
}
\value{
a dataframe
}
\description{
mapfunction()
This function maps the location of earthquake onto a leaflet map. This map function is an interactive map that the user can user and interect to see their data in a geospatial manner.
}
\examples{
\dontrun{
mapfunction(MyData)
}
}
