% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{minutesfunction}
\alias{minutesfunction}
\title{minutesfunction()
This function subset the data to the number of minutes prior to current time inserted my the user input}
\usage{
minutesfunction(x, df)
}
\arguments{
\item{x}{takes an integer. Number of minutes from current time}

\item{df}{takes a dataframe}
}
\value{
a dataframe
}
\description{
minutesfunction()
This function subset the data to the number of minutes prior to current time inserted my the user input
}
\examples{
MyData = importusgsdata(day) #importing dataframe
minutesfunction(300, MyData) #data from 300 minutes ago till current time.
}
