% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{timeconvert}
\alias{timeconvert}
\title{timeconvert()}
\usage{
timeconvert(vector)
}
\arguments{
\item{vector}{takes a string}
}
\value{
returns a POSIXlt date-time class.
}
\description{
This function is a helper function to strip the string format of the Date-Time format and convert it to POSIXlt format
}
\examples{
\dontrun{
as.POSIXlt(sapply(MyData$time,FUN=timeconvert))
}
}
