% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{weekfunction}
\alias{weekfunction}
\title{weekfunction()
This function subset the data to the number of week prior to current time inserted my the user input}
\usage{
weekfunction(x, df)
}
\arguments{
\item{x}{takes an integer. Number of wek from current time}

\item{df}{takes a dataframe}
}
\value{
a dataframe
}
\description{
weekfunction()
This function subset the data to the number of week prior to current time inserted my the user input
}
\examples{
MyData = importusgsdata(day) #importing dataframe
weekfunction(2, MyData) #data from 2 weeks ago till current time.
}
