% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{customusgslink}
\alias{customusgslink}
\title{customusgslink()}
\usage{
customusgslink(URL, df = FALSE)
}
\arguments{
\item{URL}{takes a URL from USGS website}

\item{df}{User choose to export dataframe automatically to the global environment called MyData. Defaults to FALSE.}
}
\value{
a dataframe class. Return a dataframe from the USGS link provided by the user, with parameters of time, lat, long, depth, mag, magType, place, locations and, DateTime.
}
\description{
This function imports custom URL link obtained from USGS. Input is in CSV format.
}
\examples{
MyData = customusgslink("https://earthquake.usgs.gov/earthquakes/feed/v1.0/summary/all_hour.csv", TRUE)
}
